いいねのAjaxを管理するpost.jsを作成、showに読み込ませる
