+++ app/views/posts/show.html.haml
@@ -6,3 +6,5 @@
         %span.arrow-left
 
   = render 'commons/posts', post: @post
+
+= javascript_pack_tag 'post'
